Fixed performance-tests.
Moved fixtures shared between 'functional' and 'performance' up
one level.

Fixed some problems w/ the tests themselves as well.

fixes #3095
