+++ CHANGES/3095.misc
@@ -1,0 +1,1 @@
+Fixed performance-tests.
